+++ src/debug/mod.rs
@@ -1,3 +1,0 @@
-mod display_as_debug;
-
-pub use display_as_debug::DisplayAsDebug;