+++ src/display/mod.rs
@@ -1,3 +1,0 @@
-mod debug_as_display;
-
-pub use debug_as_display::DebugAsDisplay;