+++ src/lib.rs
@@ -13,14 +13,5 @@ pub mod fmt;
 /// Helper types that implement [`Debug`] and/or [`Display`] for specialized formatting.
 pub mod types;
 
-/// Traits and types for treating a [`Display`] implementation as a [`Debug`] implementation.
-pub mod debug;
-
-/// Traits and types for treating a [`Debug`] implementation as a [`Display`] implementation.
-pub mod display;
-
-/// Traits and types for implementing [`Debug`] for [`Option`].
-pub mod option;
-
-/// Traits and types for implementing [`Debug`] for [`Result`].
-pub mod result;
+/// Wrapper types for [`Debug`] and [`Display`] format conversions.
+pub mod wrap;