+++ src/option/mod.rs
@@ -1,8 +1,0 @@
-mod opaque_option;
-mod type_name_option;
-
-pub(crate) const STR_NONE: &str = "None";
-pub(crate) const STR_SOME: &str = "Some";
-
-pub use opaque_option::OpaqueOption;
-pub use type_name_option::TypeNameOption;