+++ src/types/mod.rs
@@ -1,13 +1,9 @@
 //! Types that implement [`Debug`](core::fmt::Debug) for specialized formatting.
 
 mod opaque;
-mod opaque_list;
 mod test_value;
 mod type_name;
-mod type_name_list;
 
-pub use opaque::{OPAQUE, Opaque};
-pub use opaque_list::OpaqueList;
+pub use opaque::{OPAQUE, Opaque, OpaqueList, OpaqueMap, OpaqueSet};
 pub use test_value::TestValue;
-pub use type_name::{DisplayMode, Full, Short, TypeName};
-pub use type_name_list::TypeNameList;
+pub use type_name::{DisplayMode, Full, Short, TypeName, TypeNameList, TypeNameMap, TypeNameSet};